# contracts/openapi.yaml
openapi: 3.0.4

info:
  title: API de Prestamos
  version: "1.0.0"

paths:

  /v1/socios:
    get:
      summary: Listar socios
      responses:
        '200':
          description: Lista de socios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Socio'
        '401':
          $ref: '#/components/responses/No_Autorizado'

    post:
      summary: Registrar un socio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Socio_Input'
      responses:
        '201':
          description: Socio creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Socio'
        '400':
          $ref: '#/components/responses/Error_Generico'
        '401':
          $ref: '#/components/responses/No_Autorizado'

  /v1/socios/{socio_id}:
    get:
      summary: Consultar un socio
      parameters:
        - name: socio_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Socio encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Socio'
        '400':
          $ref: '#/components/responses/Error_Generico'
        '401':
          $ref: '#/components/responses/No_Autorizado'
        '404':
          $ref: '#/components/responses/No_Encontrado'

  /v1/prestamos:
    get:
      summary: Listar prestamos
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            type: string
          description: Filtra por coincidencia exacta. Si se omite, se listan todos; el servicio no valida una enumeración de valores.
      responses:
        '200':
          description: Lista de prestamos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prestamo'
        '401':
          $ref: '#/components/responses/No_Autorizado'

    post:
      summary: Registra un prestamo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prestamo_Input'
      responses:
        '201':
          description: Prestamo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prestamo'
        '400':
          $ref: '#/components/responses/Error_Generico'
        '401':
          $ref: '#/components/responses/No_Autorizado'
        '404':
          description: No existe el socio o el libro solicitado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: El libro solicitado no tiene ejemplares disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/Catalogo_No_Disponible'
        '502':
          $ref: '#/components/responses/Error_Catalogo'
        '504':
          $ref: '#/components/responses/Catalogo_Timeout'

  /v1/prestamos/{prestamo_id}:
    parameters:
        - name: prestamo_id
          in: path
          required: true
          schema:
            type: integer

    get:
      summary: Consulta un prestamo
      responses:
        '200':
          description: Prestamo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prestamo'
        '400':
          $ref: '#/components/responses/Error_Generico'
        '401':
          $ref: '#/components/responses/No_Autorizado'
        '404':
          $ref: '#/components/responses/No_Encontrado'
                
    delete:
      summary: Registrar la devolución de un préstamo
      description: Libera el ejemplar en Catálogo y cambia el estado del préstamo a DEVUELTO; no elimina el registro.
      responses:
        '200':
          description: Préstamo devuelto; el registro se conserva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prestamo'
        '404':
          $ref: '#/components/responses/No_Encontrado'
        '400':
          $ref: '#/components/responses/Error_Generico'
        '401':
          $ref: '#/components/responses/No_Autorizado'
        '409':
          description: El préstamo ya fue devuelto, o el ejemplar no existe o ya estaba disponible al intentar liberarlo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/Catalogo_No_Disponible'
        '502':
          $ref: '#/components/responses/Error_Catalogo'
        '504':
          $ref: '#/components/responses/Catalogo_Timeout'

  /health:
    get:
      summary: Estado del servicio y sus dependencias
      security: []
      responses:
        '200':
          description: Servicio y dependencias operativas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Una o más dependencias críticas no están disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

security:
  - ApiKeyAuth: []

components:
  schemas:
    
    # --- SOCIO ---

    Socio_Input:
      type: object
      required: [nombre, correo]
      properties:
        nombre: { type: string }
        correo: { type: string, format: email }

    Socio:
      allOf:
        - $ref: '#/components/schemas/Socio_Input'
        - type: object
          required: [id]
          properties:
            id: { type: integer }
            prestamos:
              type: array
              default: []
              items:
                $ref: '#/components/schemas/Prestamo'

    # --- PRESTAMO ---

    Prestamo_Input:
      type: object
      required: [socio_id, libro_id]
      properties:
        socio_id: { type: integer}
        libro_id: { type: integer}

    Prestamo:
      allOf:
        - $ref: '#/components/schemas/Prestamo_Input'
        - type: object
          required: [id, ejemplar_id, fecha_prestamo, fecha_limite, fecha_devolucion, estado]
          properties:
            id: { type: integer }
            ejemplar_id: { type: integer }
            fecha_prestamo: { type: string, format: date-time}
            fecha_limite: { type: string, format: date}
            fecha_devolucion: { type: string, format: date-time, nullable: true}
            estado: { type: string }
            _links:
              type: object
              description: Enlaces HATEOAS para navegación de estado
              additionalProperties:
                type: object
                properties:
                  href: { type: string }
                  method: { type: string }
    # --- ERROR ---
    Error:
      type: object
      required: [codigo, mensaje]
      properties:
        codigo: { type: string }
        mensaje: { type: string }

    DependencyStatus:
      type: object
      required: [database, catalogo_grpc, circuit_breaker]
      properties:
        database: { type: string }
        catalogo_grpc: { type: string }
        circuit_breaker: { type: string }

    HealthResponse:
      type: object
      required: [status, service, dependencies]
      properties:
        status: { type: string }
        service: { type: string }
        dependencies:
          $ref: '#/components/schemas/DependencyStatus'

  securitySchemes:

  # --- AUTENTICACION POR API KEY ---

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  responses:

    # --- ERRORES ---

    Error_Generico:
      description: Error de validación de parámetros o body; al registrar un socio también indica correo duplicado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    No_Autorizado:
      description: Falta una API Key válida o la clave enviada no es correcta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    No_Encontrado:
      description: El recurso solicitado no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Catalogo_No_Disponible:
      description: Catálogo no está disponible o el circuit breaker está abierto; la operación no se completó
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Error_Catalogo:
      description: Error no clasificado al comunicarse con Catálogo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Catalogo_Timeout:
      description: Catálogo no respondió antes de vencer el timeout
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'